# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: KIU Admission & Career Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admission/applications:
    post:
      operationId: createAdmissionApplication
      tags: [admission]
      summary: Submit a new admission application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdmissionApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listAdmissionApplications
      tags: [admission]
      summary: List all admission applications (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, under_review, accepted, rejected, waitlisted]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplicationList"
  /admission/applications/my:
    get:
      operationId: getMyAdmissionApplication
      tags: [admission]
      summary: Get the current user's admission application
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
        "404":
          description: No application found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admission/applications/{id}:
    get:
      operationId: getAdmissionApplication
      tags: [admission]
      summary: Get an admission application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAdmissionApplicationStatus
      tags: [admission]
      summary: Update application status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdmissionApplication"
  /admission/programs:
    get:
      operationId: listPrograms
      tags: [admission]
      summary: List available programs/courses
      parameters:
        - name: faculty
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [diploma, degree]
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramList"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [applicant, finalist, admin]
          default: applicant
        nationalId:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [applicant, finalist, admin]
        nationalId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
...
    UnebGrade:
      type: object
      properties:
        subject:
          type: string
        grade:
          type: string
        points:
          type: integer
      required:
        - subject
        - grade
        - points
    AdmissionApplicationRequest:
      type: object
      properties:
        programId:
          type: integer
        examLevel:
          type: string
          enum: [o_level, a_level]
        examYear:
          type: integer
        indexNumber:
          type: string
        unebGrades:
          type: array
          items:
            $ref: "#/components/schemas/UnebGrade"
        personalStatement:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum: [male, female, other]
        nationality:
          type: string
          default: Ugandan
        district:
          type: string
        nextOfKinName:
          type: string
        nextOfKinPhone:
          type: string
        nextOfKinRelationship:
          type: string
      required:
        - programId
        - examLevel
        - examYear
        - indexNumber
        - unebGrades
        - dateOfBirth
        - gender
    AdmissionApplication:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        programId:
          type: integer
        program:
          $ref: "#/components/schemas/Program"
        status:
          type: string
          enum: [pending, under_review, accepted, rejected, waitlisted]
        examLevel:
          type: string
          enum: [o_level, a_level]
        examYear:
          type: integer
        indexNumber:
          type: string
        unebGrades:
          type: array
          items:
            $ref: "#/components/schemas/UnebGrade"
        personalStatement:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
        nationality:
          type: string
        district:
          type: string
        nextOfKinName:
          type: string
        nextOfKinPhone:
          type: string
        nextOfKinRelationship:
          type: string
        adminNotes:
          type: string
        applicationNumber:
          type: string
        submittedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        applicantName:
          type: string
        applicantEmail:
          type: string
      required:
        - id
        - userId
        - programId
        - status
        - examLevel
        - examYear
        - indexNumber
        - unebGrades
        - applicationNumber
    AdmissionApplicationList:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/AdmissionApplication"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - applications
        - total
        - page
        - limit
    UpdateApplicationStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, under_review, accepted, rejected, waitlisted]
        adminNotes:
          type: string
      required:
        - status
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        faculty:
          type: string
        department:
          type: string
        level:
          type: string
          enum: [diploma, degree]
        duration:
          type: string
        description:
          type: string
        entryRequirements:
          type: string
        minOlevelPoints:
          type: integer
        minAlevelPoints:
          type: integer
        availableSlots:
          type: integer
      required:
        - id
        - name
        - code
        - faculty
        - level
    ProgramList:
      type: object
      properties:
        programs:
          type: array
          items:
            $ref: "#/components/schemas/Program"
        total:
          type: integer
      required:
        - programs
        - total
...