# lib/api-zod/src/generated/types/programList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KIU Admission & Career Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Program } from "./program";

export interface ProgramList {
  programs: Program[];
  total: number;
}